requirements.txt
Python>=3.8.1
Django>=3.0.2
sqlparse>=0.3.0
djangorestframework>=*
multiselectfield>=*
corsheaders>=*
mysqlclient>=*